requirements.txt
torch>=1.9.0
torchvision>=0.10.0
requests>=2.25.0
pandas>=1.3.0
Pillow>=8.0.0
albumentations>=1.0.0
tensorboard>=2.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
timm>=1.0.10
matplotlib>=3.5.0
seaborn>=0.12.0
joblib>=1.2.0